fix(CustomEndpoints): further refactor
